feat!: disable es6-like string interpolation ${} when compiling templates
This patch disables using ${} in the html templates. This is mostly
to avoid conflicts with the provided template that might use ${} in
inline javascript natively.

This patch also reverts the changes of #46.
lodash.template is now always called. this is useful for when you want
to use `process.env`.

Therefore, this change is breaking, as it removes ${}, but also always
calls lodash again.

If this ever will be a problem, we can also add a configuration option
to enable ${} manually again.